ANDDEP-1207 extracted FragmentNavigator, TabFragmentNavigator interfaces
