+++ core-navigation/lib-core-navigation/src/main/java/ru/surfstudio/android/core/ui/navigation/fragment/FragmentNavigatorImpl.java
@@ -17,40 +17,24 @@ package ru.surfstudio.android.core.ui.navigation.fragment;
 
 
 import androidx.annotation.IdRes;
-import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
 import ru.surfstudio.android.core.ui.FragmentContainer;
-import ru.surfstudio.android.core.ui.navigation.Navigator;
 import ru.surfstudio.android.core.ui.navigation.fragment.route.FragmentRoute;
 import ru.surfstudio.android.core.ui.provider.ActivityProvider;
 
-import static android.app.FragmentTransaction.TRANSIT_FRAGMENT_CLOSE;
-import static android.app.FragmentTransaction.TRANSIT_FRAGMENT_FADE;
-import static android.app.FragmentTransaction.TRANSIT_FRAGMENT_OPEN;
-import static android.app.FragmentTransaction.TRANSIT_NONE;
 
-/**
- * позволяет осуществлять навигацияю между фрагментами
- */
-public class FragmentNavigator implements Navigator {
+public class FragmentNavigatorImpl implements FragmentNavigator {
     protected final ActivityProvider activityProvider;
 
-    @IntDef({TRANSIT_NONE, TRANSIT_FRAGMENT_OPEN, TRANSIT_FRAGMENT_CLOSE, TRANSIT_FRAGMENT_FADE})
-    @Retention(RetentionPolicy.SOURCE)
-    private @interface Transit {
-    }
-
-    public FragmentNavigator(ActivityProvider activityProvider) {
+    public FragmentNavigatorImpl(ActivityProvider activityProvider) {
         this.activityProvider = activityProvider;
     }
 
+    @Override
     public void add(FragmentRoute route, boolean stackable, @Transit int transition) {
         int viewContainerId = getViewContainerIdOrThrow();
         FragmentManager fragmentManager = getFragmentManager();
@@ -66,6 +50,7 @@ public class FragmentNavigator implements Navigator {
         fragmentTransaction.commit();
     }
 
+    @Override
     public void replace(FragmentRoute route, boolean stackable, @Transit int transition) {
         int viewContainerId = getViewContainerIdOrThrow();
         FragmentManager fragmentManager = getFragmentManager();
@@ -81,9 +66,7 @@ public class FragmentNavigator implements Navigator {
         fragmentTransaction.commit();
     }
 
-    /**
-     * @return возвращает {@code true} если фрагмент был удален успешно
-     */
+    @Override
     public boolean remove(FragmentRoute route, @Transit int transition) {
         FragmentManager fragmentManager = getFragmentManager();
         fragmentManager.executePendingTransactions();
@@ -101,23 +84,17 @@ public class FragmentNavigator implements Navigator {
         return true;
     }
 
-    /**
-     * @return возвращает {@code true} если фрагмент успешно отобразился
-     */
+    @Override
     public boolean show(FragmentRoute route, @Transit int transition) {
         return toggleVisibility(route, true, transition);
     }
 
-    /**
-     * @return возвращает {@code true} если фрагмент был скрыт успешно
-     */
+    @Override
     public boolean hide(FragmentRoute route, @Transit int transition) {
         return toggleVisibility(route, false, transition);
     }
 
-    /**
-     * @return возвращает {@code true} если какой-либо фрагмент верхнего уровня был удален из стека
-     */
+    @Override
     public boolean popBackStack() {
         FragmentManager fragmentManager = getFragmentManager();
         fragmentManager.executePendingTransactions();
@@ -125,20 +102,7 @@ public class FragmentNavigator implements Navigator {
         return fragmentManager.popBackStackImmediate();
     }
 
-    /**
-     * Очищает стек фрагментов до роута.
-     * Пример:
-     * Фрагменты А, Б, С, Д добавлены в стек
-     * popBackStack(Б, true) очистит стек до Б включительно
-     * то есть, останется в стеке только А
-     * или
-     * popBackStack(Б, false) в стеке останется А и Б,
-     * Б не удаляется из стека
-     *
-     * @param route     очистка до уровня роута
-     * @param inclusive удалить стек включая и роут
-     * @return возвращает {@code true} если фрагмент(ы) был(и) удален(ы) из стека
-     */
+    @Override
     public boolean popBackStack(@NonNull FragmentRoute route, boolean inclusive) {
         FragmentManager fragmentManager = getFragmentManager();
         fragmentManager.executePendingTransactions();
@@ -173,11 +137,7 @@ public class FragmentNavigator implements Navigator {
                 inclusive ? FragmentManager.POP_BACK_STACK_INCLUSIVE : 0);
     }
 
-    /**
-     * Очистка бэкстека
-     *
-     * @return true если успешно
-     */
+    @Override
     public boolean clearBackStack() {
         FragmentManager fragmentManager = getFragmentManager();
         fragmentManager.executePendingTransactions();
