+++ dagger-scope/sample/src/main/java/ru/surfstudio/android/custom_scope_sample/ui/base/dagger/activity/ActivityModule.kt
@@ -7,6 +7,7 @@ import ru.surfstudio.android.core.ui.event.ScreenEventDelegateManager
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigator
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForActivity
 import ru.surfstudio.android.core.ui.navigation.fragment.FragmentNavigator
+import ru.surfstudio.android.core.ui.navigation.fragment.FragmentNavigatorImpl
 import ru.surfstudio.android.core.ui.permission.PermissionManager
 import ru.surfstudio.android.core.ui.permission.PermissionManagerForActivity
 import ru.surfstudio.android.core.ui.provider.ActivityProvider
@@ -91,6 +92,6 @@ class ActivityModule(private val persistentScope: ActivityPersistentScope) {
     @Provides
     @PerActivity
     internal fun provideFragmentNavigator(activityProvider: ActivityProvider): FragmentNavigator {
-        return FragmentNavigator(activityProvider)
+        return FragmentNavigatorImpl(activityProvider)
     }
 }
