+++ template/base_feature/src/main/java/ru/surfstudio/standard/ui/activity/di/ActivityModule.kt
@@ -9,7 +9,9 @@ import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavig
 import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavigatorForActivity
 import ru.surfstudio.android.core.ui.navigation.feature.installer.SplitFeatureInstaller
 import ru.surfstudio.android.core.ui.navigation.fragment.FragmentNavigator
+import ru.surfstudio.android.core.ui.navigation.fragment.FragmentNavigatorImpl
 import ru.surfstudio.android.core.ui.navigation.fragment.tabfragment.TabFragmentNavigator
+import ru.surfstudio.android.core.ui.navigation.fragment.tabfragment.TabFragmentNavigatorImpl
 import ru.surfstudio.android.core.ui.permission.PermissionManager
 import ru.surfstudio.android.core.ui.permission.PermissionManagerForActivity
 import ru.surfstudio.android.core.ui.provider.ActivityProvider
@@ -90,7 +92,7 @@ class ActivityModule(private val persistentScope: ActivityPersistentScope) {
     @Provides
     @PerActivity
     internal fun provideFragmentNavigator(activityProvider: ActivityProvider): FragmentNavigator {
-        return FragmentNavigator(activityProvider)
+        return FragmentNavigatorImpl(activityProvider)
     }
 
     @Provides
@@ -124,10 +126,11 @@ class ActivityModule(private val persistentScope: ActivityPersistentScope) {
     @Provides
     @PerActivity
     internal fun provideTabFragmentNavigator(
-            activityProvider: ActivityProvider,
-            eventDelegateManager: ScreenEventDelegateManager
+        activityProvider: ActivityProvider,
+        fragmentNavigator: FragmentNavigator,
+        eventDelegateManager: ScreenEventDelegateManager
     ): TabFragmentNavigator {
-        return TabFragmentNavigator(activityProvider, eventDelegateManager)
+        return TabFragmentNavigatorImpl(activityProvider, fragmentNavigator, eventDelegateManager)
     }
 
     @Provides
